git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/ERM.BL/Dev/Features/ERM-2842-DateTime/BLFlex;C25834

+++ Source/ApplicationServices/2Gis.Erm.BLFlex.Operations.Global/Shared/Formatter/INumberToWordsConverter.cs
@@ -2,6 +2,6 @@
 {
     public interface INumberToWordsConverter
     {
-        string Convert(int number);
+        string Convert(long number);
     }
 }
